+++ docs/pages/changelog.rst
@@ -45,6 +45,7 @@ Fixed
 - Fixed estimation of required storage to keep device restore files within the specified maximum storage.
 - Fixed crash due to inconsistent channel values and channel background. Instead showing warning and restoring consistency.
 - Fixed issue on |omnicontrol| :ref:`sec:omnicontrol` with loading values with mixed channel data type. Make sure that the default channel data type is suitable for exporting and restoring data of all other types. E.g. if you mix float (pump speed) and exp (pressure), use exp as the default data type.
+- Fixed loss of history after loading configuration by storing before and loading after loadConfiguration.
 
 Developer Notes
 ~~~~~~~~~~~~~~~